fix(setLocation): post new location notification twice (workaround).
The new location notification is posted twice as a workaround to overcome the issue of not notifying the app for the new location that was set on the first attempt after setting a new location permissions.

We post the new location notifications twice, and the location is set only on the second attempt.

The mentioned bug also happens when setting the location directly through the Simulator app (without AppleSimUtils).